Abstract Question into a new model
-(feature) Abstract Question into a new model
-(feature) Add Question routes
-(feature) Assign Questions and Exams ownership at creation
-(feature) Check for ownership before delete or update
-(bug) Fix sendStatus deprecation

+++ controllers/exam.controller.js
@@ -10,13 +10,14 @@ exports.index = (req, res, next) => {
 exports.show = function(req, res, next) {
   Exam.getExam(req.params.id, (err, exam) => {
     if (err) return next(err)
-    if (!exam) return res.send(401)
+    if (!exam) return res.sendStatus(404)
     res.json(exam)
   })
 }
 
 exports.create = function(req, res, next) {
   //TODO validate req.body
+  req.body._owner = req.user._id
   Exam.createExam(req.body, err => {
     if (err) {
       res.send(err)
@@ -30,19 +31,29 @@ exports.create = function(req, res, next) {
 }
 
 exports.update = function(req, res, next) {
-  Exam.updateExam(req.params.id, req.body, err => {
-    //TODO validate req.body
-    if (err) res.send(err)
-    res.json({
-      success: true,
-      message: 'Exam Updated'
+  Exam.getExam(req.params.id, (err, exam) => {
+    if (err) return next(err)
+    if (!exam) return res.sendStatus(404)
+    if (exam._owner !== req.user._id) return res.sendStatus(401)
+    Exam.updateExam(req.params.id, req.body, err => {
+      //TODO validate req.body
+      if (err) res.send(err)
+      res.json({
+        success: true,
+        message: 'Exam Updated'
+      })
     })
   })
 }
 
 exports.destroy = function(req, res, next) {
-  Exam.deleteExam(req.params.id, err => {
-    if (err) return res.send(500, err)
-    return res.send(204)
+  Exam.getExam(req.params.id, (err, exam) => {
+    if (err) return next(err)
+    if (!exam) return res.sendStatus(404)
+    if (exam._owner !== req.user._id) return res.sendStatus(401)
+    Exam.deleteExam(req.params.id, err => {
+      if (err) return res.send(500, err)
+      return res.sendStatus(204)
+    })
   })
 }
